pyproject.toml
[project]
name = "telegramm-shorts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.5",
    "yt-dlp>=2025.12.8",
]
